chore: rename data-summary component
