+++ apps/picsa-apps/dashboard/src/app/modules/climate/pages/station-details/components/data-summary/data-summary.scss
