+++ apps/picsa-apps/dashboard/src/app/modules/climate/pages/station-details/components/data-summary/data-summary.ts
@@ -19,7 +19,7 @@ interface IRainfallSummary {
 }
 
 @Component({
-  selector: 'dashboard-climate-rainfall-summary',
+  selector: 'dashboard-climate-data-summary',
   templateUrl: './rainfall-summary.html',
   imports: [DashboardClimateApiStatusComponent, PicsaDataTableComponent],
   styleUrl: './rainfall-summary.scss',
